refactor: file naming and interface file for each feature
