+++ apps/datacollector/lib/Display/display_interface.h
@@ -1,0 +1,22 @@
+// ** DISPLAY_CONTROLLER **
+#define LED_BUILTIN GPIO_NUM_25
+#define PIN_35 GPIO_NUM_35
+
+#define SCREEN_WIDTH 128
+#define SCREEN_HEIGHT 64
+#define SCREEN_H_CENTER SCREEN_WIDTH / 2
+#define SCREEN_V_CENTER SCREEN_HEIGHT / 2
+
+#define SCREEN_PADDING 7
+#define SCREEN_TOP_PADDING SCREEN_PADDING
+#define SCREEN_BOTTOM_PADDING SCREEN_HEIGHT - SCREEN_PADDING
+#define SCREEN_TOP_RIGHT_CURSOR_POSITION SCREEN_WIDTH
+#define SCREEN_BOTTOM_RIGHT_CURSOR_POSITION SCREEN_WIDTH
+#define SCREEN_TOP_LEFT_CURSOR_POSITION 0
+#define SCREEN_BOTTOM_LEFT_CURSOR_POSITION 0
+
+#define TEXT_START_LINE 2
+#define TEXT_END_LINE 7
+
+#define OLED_RESET -1
+#define SCREEN_ADDRESS 0x3D
