+++ apps/datacollector/lib/Input/input_controller.cpp
@@ -1,7 +1,4 @@
-#include <input_controller.h>
-
-#define LED_BUILTIN GPIO_NUM_25
-#define BTN_CONFIRM GPIO_NUM_35
+#include "input_controller.h"
 
 InputController *InputController::instance = nullptr;
