+++ apps/datacollector/lib/Input/input_interface.h
@@ -1,0 +1,3 @@
+// ** INPUT_CONTROLLER **
+#define LED_BUILTIN GPIO_NUM_25
+#define BTN_CONFIRM GPIO_NUM_35
