+++ apps/datacollector/lib/Sensor/sensor_controller.cpp
@@ -2,7 +2,6 @@
 
 SensorController *SensorController::instance = nullptr;
 
-
 SensorController::SensorController(){
     head = NULL;
     tail = NULL;
