+++ apps/datacollector/lib/Utils/file_handler.h
@@ -1,6 +1,6 @@
 #pragma once
-#include "SPIFFS.h"
 #include <ArduinoJson.h>
+#include "SPIFFS.h"
 
 class FileHandler
 {
