+++ apps/datacollector/lib/api/ontop_client.cpp
@@ -65,3 +65,8 @@ int OnTopClient::initDevice(){
 
     return statusCode;
 }
+
+int OnTopClient::syncDevice()
+{
+
+}
